GD-60: Allow to run a single parameterized test from the inspector
# Why
By default, you can execute only one selected test or test suite. For parameterized tests, the inspector is populated with all parameterized tests when the test is executed. If a test fails, you have to re-run the entire test case, but it would be nice if you could select one of the parameterized test runs and re-run it as a single case.

![image](https://user-images.githubusercontent.com/347037/208474744-c29e65a6-74cf-43dd-883b-d9cf0cd0eb58.png)


# What
Implemented to run a selected parameterized test case via the inspector context menu.

![image](https://user-images.githubusercontent.com/347037/208474980-482b9954-81eb-4702-8e82-7974027f8811.png)