GD-123: Change to rendermode opengl for c-pr
# Why
In the current ci-pr WF, the Godot engine runs in headless mode due to missing display and render drivers.
Using headless mode only works for tests without UI interactions and will not work with the GdUnit4 SceneRunner.
The SceneRunner uses Input.warp_mouse() and Input.parse_input_event() which are not supported in headless mode.

# What
- exit with error 103 if starting the cmd tool with option `--headless`
- remove the headless mode from ci-pr unit test run
- download and install opengl driver if need
- run tests now with option `--display-driver x11 --rendering-driver opengl3 --screen 0`