Replace download host for Godot with GitHub (#227)
# Why
TuxFamily is currently down (again) and also slow. Because TuxFamily is
down, the latest Godot release cannot be uploaded and is therefore not
available to this action.

Downloading from GitHub is faster, more reliable and therefore reduces
GitHub Action runtime


# What
This changes the default download host from TuxFamily to GitHub.