GD-221: Fix GdUnitInspector button bar issues when a scene is open
# Why
The button bar did not adjust the execution state when the test was running and a scene was open in the editor.

# What
There was an error in handling the current status of the test execution.
We perform a periodic `check_test_run_stopped_manually` to ensure that we eventually stop when the test is manually stopped.
With this check, we test for `_is_running==true`, but the status was set to true too early.
- set status _is_running is now only set to true when the testrunner is started and not before.
- fix warning `unused_parameter` at SceneRunner#set_property