GD-258: Allow to use argument matcher for arguments on mocks
# Why
Currently, all mocks have default return values for functions and can be explicitly overridden.
This makes it complicated to define a return value for a function with arguments that covers multiple argument variations. You are forced to define a do_return() on the mock for each possible argument variation.

# What
- allow now to use any argument matcher to configure do_return values on a mocked functions.
instead of defining for each possible argument variation you can now use `any()` or the typed any machtchers like `any_int()`