GD-389: The inheritance of a base test class leads to an incorrect test case list.
# Why
https://github.com/MikeSchulze/gdUnit4/issues/389

# What
When the scanner is building the function descriptors, it was using the list of tests and adding the `before` function to scan. The list to scan was not a copy and results into side effect of the original test list.
Fixed by using a copy before adding the `before` function
Extended the existing test coverage for test suite inheritance to include the `before` on the base test suite.