Simplify `.gitattributes` (#424)
# Why
I can't tell for sure but it seems like `.runsettings*` and `.gdlintrc`
paths were included in the release by mistake since the `.gitattributes`
is an excluded list.

The current `.gitattributes` structure looks difficult to maintain. If
`.runsettings*` and/or `.gdlintrc` should be kept in the final release,
this PR can easily be updated to include them explicitly.

# What
I flipped the `.gitattributes` around to be an includes list by
excluding everything first. This approach should keep it more
maintainable as you may add other files to the project later that are
not needed in the final release.

This is testable by running this command against this branch:
`git archive --format=tar improve-gitattribute | gzip >
<replace-this-with-a-path>`