Improve settings UI (#534)
Hi, this PR improves the settings dialogue UI, making it significantly
more compact, and IMHO, also easier to understand and nicer to look at.

Before:

![gdunit-before](https://github.com/MikeSchulze/gdUnit4/assets/179487/cae07ac3-3065-4f41-802d-ecde475a95a7)

After:

![gdunit-after](https://github.com/MikeSchulze/gdUnit4/assets/179487/0265e7e3-e75c-46cc-9da5-892940597bcb)


The changes are split into three commits

1. Remove the value sets of settings from the user-facing help strings.
Fixes #539
2. Improve the wording of the help strings themselves (more detailed
description in the commit message). Fixes #540
3. Improved sizing logic for the grid views displaying the settings,
meaning less space is wasted, and huge amounts of unused space don't
make the scrollbars huge unnecessarily. Not every tab in the dialogue is
currently able to fit in without scrolling on my laptop, but some of
them are, and it's generally much nicer. Fixes #541


I did not make any changes to the settings themselves, everything is as
it was before and split across tabs in the same way. I did a full test
suite run locally, and everything works aside from the C# tests, which I
did not run, as I'm not using Godot Mono.

---------

Co-authored-by: Nullpointer <mikeschulze.lpz@gmail.com>