Added support for Godot4.4-dev5 (#600)
# Why
With Godot-4.4-dev5 the uid files were introduced.


# What
we have to explicitly exclude them from the repository. 
The official Godot release nodes says we have to include them, but this
is not true.

https://godotengine.org/article/dev-snapshot-godot-4-4-dev-5/
```
Note that, unlike metadata files, .uid files are strictly for the editor; the information is migrated to the uid database on export. Despite this, users using version control software should add these files, as they’re required to properly sync data (similar to .import files). For more information, see (GH-97352).
```

- fix failing tests according to new float vector string formattings.
- fix failing patcher test by exclude uid files.