Fix crash and invalid state on cmd tool when no tests found (#832)
# Why
When using the cmd tool to run a test where not exists, the tool crashes
with an error.
```
Debugger Break, Reason: 'Attempt to call function 'execute_startup' in base 'previously freed' on a null instance.'
*Frame 0 - res://addons/gdUnit4/src/core/runners/GdUnitTestSessionRunner.gd:120 in function '_process'
Enter "help" for assistance.
debug>
```

# What
Do exit the program with the correct state by return to avoid to set
state to RUN