GD-878: Fixes verify a signal is NOT emitted twice by using argument matchers (#896)
# Why
If a user wants to test whether a signal is emitted and then check that
it is not emitted again, this will fail if the Any matcher is used as
the signal parameter.

# What
- Add missing test coverage
- We delete the matching received signal argument using
`GdObjects.equals`. Using array `erase` uses object reference equality,
which does not work for argument matchers.

(cherry picked from commit 49698786f59d920593343b0ecafe4fea2d430104)