GD-949: Allow underscore prefix for test case arguments to avoid unused argument warnings
# Why
As test parameters is only meant to be declared (and not used), the GDScript compiler warns it's declared but not used.
To prevent this warning the user must set `@warning_ignore("unused_parameter")` on the function signature.

# What
Allowing to use test function arguments with prefix underscore to prevent spamming warnings
`_test_parameters`, `_timeout`, `_do_skip`, `skip_reason`