GD-1010: C# When the project is recompiled, it causes a duplicate key exception in ScriptTypeBiMap
# Why
When the project is recompiled, it causes a `duplicate key exception in ScriptTypeBiMap` because the class `GdUnit4CSharpApi` is inherits form `GdUnit4NetApiGodotBridge`. The class `GdUnit4NetApiGodotBridge` is shipped by the gdUnit4.api nuget package and contains non Godot conform class/objects where is causing into this issue.

# What
Instead of inherit we now delegate the method calls to the static methods provided by the `GdUnit4NetApiGodotBridge`